implement new home page card ui
